Support AWS and GCP special user scenarios
CREATE SUBSCRIPTION requires that the user is superuser, which is why require the user for source and target
url/con string to be superuser. However, AWS and GCP don't have a superuser role but expose dedicated roles
that act as super user (rds_superuser and cloudsqlsuperuser) and can perform create subscription.

This PR now supports passing such user role during bootstrap and config check. We then accordingly ensure that the
internal user we create for replication has been granted the passed in special_user_role accordingly. Its a little
hard to write spec for since we still need grantee role to be superuser.

We have also relaxed permissions during config check. If a special_user_role is passed, we just make sure that user
on the source and target db is a member of the special_user_role. If special_user_role is not passed, we expect
the user to be a superuser.