Don't attempt to drop and recreate unique indices
Droping unique indices is not possible w/o having to drop and recreate the constraint.

We could in future consider doing that if there are significant performance gains. Otherwise,
just don't attempt to drop/recreate it.

We'd still gain on the performance improvement from having dropped other indices before copy