fixed typo, removed spaces from tables list example
When I executed the migration for only a small subset of tables, I did notice that it would error out if there were spaces in the list of tables.  Just updating the example to reflect no spaces are needed.  And I saw the track-ddl command was missing a dash at the front.